feat: add ItemData hierarchi with WeaponData, Rarity and WeaponType

+++ src/Presentation/Strategies/Attack/MeleeAttack.cs
@@ -4,8 +4,8 @@ namespace CombatLab.Presentation.Strategies.Attack;
 
 public class MeleeAttack : IAttackStrategy
 {
-    private int _damage;
-    public MeleeAttack(int damage) {_damage = damage;}
+    private float _damage;
+    public MeleeAttack(float damage) {_damage = damage;}
     public void Execute(IAttacker attacker, IDamageable target)
     {
         target.TakeDamage(_damage, attacker.GlobalPosition);
